+++ CHANGES.md
@@ -1,16 +1,27 @@
 # Change log
 
-### Version 0.2.5 — xx/xx/2016
+### Version 0.3.0
+
+- Multi-color SVG support
+
+For compatibility:
+
+- `sax-classes()` mixin replaces `saxicon-classes()`
+- `sax-width()` function replaces `siconax-width()`
+- `sax-height()` function replaces `saiconx-height()`
+- `sax` function replaces `saxicon-background()`
+
+### Version 0.2.5 — xx/xx/2017
 
 - Integration with Travis CI
 - Fixed documentation
 - Updated license
 
-### Version 0.2.4 — 01/12/2016
+### Version 0.2.4 — 01/12/2017
 
 - Updated dependant version of `grunt`, `grunt-contrib-clean`, `grunt-contrib-jshint`, `grunt-contrib-nodeunit`. Namely, `grunt-contrib-jshint@0.9.2` still had the DoS vulnerability.
 
-### Version 0.2.3 — 30/11/2016
+### Version 0.2.3 — 30/11/2017
 
 - `scss.output` option is now `scss`.
 - Updated dimension handling to first use `width` & `height` attributes. If unavailable, attempt to parse these values from the `viewBox` attribute.
@@ -22,6 +33,6 @@
 
 Versions 0.2.0—0.2.3 are identical. Only documentation has changed.
 
-### Version 0.1.0 — 16/05/2016
+### Version 0.1.0 — 16/05/2017
 
 - First release.